backend/requirements.txt
# Web Framework
Flask==2.3.3
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5

# Veritabanı
psycopg2-binary==2.9.7
SQLAlchemy==2.0.21

# Dosya İşleme
PyPDF2==3.0.1
pdfplumber==0.9.0
python-docx==0.8.11
openpyxl==3.1.2
pandas==2.1.1

# OCR
pytesseract==0.3.10
Pillow==10.0.1

# AI ve NLP
langchain==0.0.325
openai==0.28.1
chromadb==0.4.15
sentence-transformers==2.2.2

# Güvenlik
bcrypt==4.0.1
PyJWT==2.8.0

# Yardımcı Kütüphaneler
python-dotenv==1.0.0
requests==2.31.0
werkzeug==2.3.7


